+++ include/stereo_depth_cl/Frame.h
@@ -17,8 +17,8 @@ public:
     unsigned long long int timestamp;
     cv::Mat classes_original_size;
     cv::Mat probs_original_size;
-    Eigen::Matrix3d K;
-    Eigen::Affine3d tf_cam_world;
+    Eigen::Matrix3f K;
+    Eigen::Affine3f tf_cam_world;
 
     int cam_id; //id of the camera depending on how many cameras we have (it gos from 0 to 1 in the case of stereo)
     int frame_id;
